보드 -> 메치보드 합침
