+++ src/main/java/com/sportsmania/swith/domain/BoardJjimVO.java
@@ -1,0 +1,13 @@
+package com.sportsmania.swith.domain;
+
+import lombok.*;
+
+@Getter
+@Builder
+@AllArgsConstructor
+@NoArgsConstructor
+public class BoardJjimVO {
+    private int wish_no;
+    private String userid;
+    private int board_no;
+}
