+++ src/main/java/com/sportsmania/swith/domain/MatchBoardVO.java
@@ -15,10 +15,12 @@ public class MatchBoardVO {
     private int board_no;
     private String board_writer;
     private String exercise;
+
     private String startdate;
     private String enddate;
     private String region;
     private int recruitsNum;
+
     private String title;
     private String content;
     private String b_category;
