+++ src/main/java/com/sportsmania/swith/dto/BoardJjimDTO.java
@@ -1,0 +1,16 @@
+package com.sportsmania.swith.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@Builder
+@AllArgsConstructor
+@NoArgsConstructor
+public class BoardJjimDTO {
+    private int wish_no;
+    private String userid;
+    private int board_no;
+}
