+++ src/main/java/com/sportsmania/swith/mapper/MatchBoardMapper.java
@@ -1,6 +1,8 @@
 package com.sportsmania.swith.mapper;
 
 import com.sportsmania.swith.domain.MatchBoardVO;
+import com.sportsmania.swith.dto.MatchBoardDTO;
+import com.sportsmania.swith.dto.PageRequestDTO;
 import org.apache.ibatis.annotations.Mapper;
 
 import java.util.List;
@@ -18,6 +20,8 @@ public interface MatchBoardMapper {
 
     void delete(int board_no);
 
+    List<MatchBoardDTO> selectList(PageRequestDTO pageRequestDTO);
+    int getCount(PageRequestDTO pageRequestDTO);
