--- src/main/java/com/sportsmania/swith/service/BoardService.java
@@ -1,13 +1,0 @@
-package com.sportsmania.swith.service;
-
-import com.sportsmania.swith.dto.BoardDTO;
-import com.sportsmania.swith.dto.PageRequestDTO;
-import com.sportsmania.swith.dto.PageResponseDTO;
-import org.springframework.stereotype.Service;
-
-
-public interface BoardService {
-    void register(BoardDTO boardDTO);
-    PageResponseDTO<BoardDTO> getList(PageRequestDTO pageRequestDTO);
-
-}
