+++ src/main/java/com/sportsmania/swith/service/MatchBoardService.java
@@ -1,6 +1,8 @@
 package com.sportsmania.swith.service;
 
 import com.sportsmania.swith.dto.MatchBoardDTO;
+import com.sportsmania.swith.dto.PageRequestDTO;
+import com.sportsmania.swith.dto.PageResponseDTO;
 
 import java.util.List;
 
@@ -16,5 +18,7 @@ public interface MatchBoardService {
     void remove(int board_no);
 
 
+
+    PageResponseDTO<MatchBoardDTO> getList(PageRequestDTO pageRequestDTO);
 
 }
