+++ src/main/resources/application.properties
@@ -11,3 +11,9 @@ mybatis.mapper-locations=classpath:mapper/**/**.xml
 spring.mvc.pathmatch.matching-strategy=ant_path_matcher
 
 jwt.secret=DSADWQeqwe23213214DASDW
+
+spring.servlet.multipart.max-request-size=30MB
+spring.servlet.multipart.max-file-size=10MB
+spring.servlet.multipart.location=C:\\upload
+
+com.sportsmania.upload.path=C:\\upload
