+++ src/main/resources/mappers/BoardMapper.xml
@@ -6,11 +6,11 @@
     <select id="getTime" resultType="string">
         select now()
     </select>
-    <insert id="insert" parameterType="com.sportsmania.swith.dto.BoardDTO">
+    <insert id="insert" parameterType="com.sportsmania.swith.dto.MatchBoardDTO">
         insert into matchboard(board_writer,exercise,startdate,enddate,region,recruitsNum,title,content,b_category,sportsfacility,supporttype)
         values(#{board_writer},#{exercise},#{startdate},#{enddate},#{region},#{recruitsNum},#{title},#{content},#{b_category},#{sportsfacility},#{supporttype})
         </insert>
-    <select id="selectList" resultType="com.sportsmania.swith.dto.BoardDTO">
+    <select id="selectList" resultType="com.sportsmania.swith.dto.MatchBoardDTO">
         select board_no
         , board_writer as board_writer
         , exercise as exercise
@@ -44,5 +44,8 @@
     </select>
     <select id="getCount" resultType="int">
         select count(board_no) from matchboard
+    </select>
+    <select id="selectOne" resultType="com.sportsmania.swith.domain.BoardVO">
+        select * from matchboard where board_no = #{board_no}
     </select>
 </mapper>
