+++ src/main/resources/mappers/MatchBoard.xml
@@ -16,7 +16,41 @@
         select * from matchboard where board_no = #{board_no}
     </select>
 
-
+    <select id="selectList" resultType="com.sportsmania.swith.dto.MatchBoardDTO">
+        select board_no
+        , board_writer as board_writer
+        , exercise as exercise
+        , startdate as startdate
+        , enddate as enddate
+        , (DATE_SUB(startdate,interval 1 hour)) as deadline
+        , region as region
+        , recruitsNum as recruitsNum
+        , title as title
+        , content as content
+        , b_category as b_category
+        , finished as finished
+        , regdate as regdate
+        , clicks as clicks
+        , sportsfacility as sportsfacility
+        , supporttype as supporttype
+        from matchboard
+        <where>
+            <if test="types != null and types.length > 0">
+                <foreach collection="types" item="type" open="(" close=")" separator="OR">
+                    <if test="type == 't'.toString()">
+                        title like concat('%',#{keyword},'%')
+                    </if>
+                    <if test="type == 'w'.toString()">
+                        writer like concat('%',#{keyword},'%')
+                    </if>
+                </foreach>
+            </if>
+        </where>
+        order by board_no desc limit #{skip},#{size}
+    </select>
+    <select id="getCount" resultType="int">
+        select count(board_no) from matchboard
+    </select>
